+++ changelog.md
@@ -1,0 +1,10 @@
+# Changelog for in-progress branch
+
+## 04/06/2021 07:48
+Returning to this project after a long break, which is largely because I had a lot of other learning to do before going much further with this one. I have now started the freeCodeCamp Tensorflow course, so I feel like I will actually be able to start making progress with this one soon. Changes to be implemented are listed below:
+
+- Switch from MySQL to SQLite
+    - MySQL is being difficult on Raspberry Pi w/ USB storage, so to simplify things, I will be switching to a SQLite database. I will likely also not at first be using all of the extra data I had accumulated the first time, and will be relying mainly on yfinance.
+- The data gathering script will run ad infinitum, so the project will be ready whenever I am able to actually start training. 
+
+New changes are being drafted in the sketches.ipynb notebook.
