--- misc_notes.txt
@@ -1,2 +1,0 @@
-1/18/21 17:58
-going to try running the yf get history file; unsure how it will work. committing and pushing to github prior to working. probably done for the night.
